Stop using parse_media_id
