+++ synapse/rest/media/v1/_base.py
@@ -23,7 +23,7 @@ from six.moves import urllib
 from twisted.internet import defer
 from twisted.protocols.basic import FileSender
 
-from synapse.api.errors import Codes, cs_error
+from synapse.api.errors import Codes, SynapseError, cs_error
 from synapse.http.server import finish_request, respond_with_json
 from synapse.util import logcontext
 from synapse.util.stringutils import is_ascii
@@ -350,3 +350,26 @@ def _parseparam(s):
         f = s[:end]
         yield f.strip()
         s = s[end:]
+
+
+def parse_media_id(request):
+    try:
+        # This allows users to append e.g. /test.png to the URL. Useful for
+        # clients that parse the URL to see content type.
+        server_name, media_id = request.postpath[:2]
+
+        if isinstance(server_name, bytes):
+            server_name = server_name.decode('utf-8')
+            media_id = media_id.decode('utf8')
+
+        file_name = None
+        if len(request.postpath) > 2:
+            try:
+                file_name = urllib.parse.unquote(request.postpath[-1].decode("utf-8"))
+            except UnicodeDecodeError:
+                pass
+        return server_name, media_id, file_name
+    except Exception:
+        raise SynapseError(
+            404, "Invalid media id token %r" % (request.postpath,), Codes.UNKNOWN
+        )
